src/core/handlers/admin_handlers.py
"""
Admin command handlers.

This module contains all the admin command handlers for the bot.
"""

import logging
from telegram import Update
from telegram.ext import ContextTypes


class AdminHandlers:
    """Handles all admin commands and interactions."""
    
    def __init__(self):
        """Initialize admin handlers."""
        self.logger = logging.getLogger(__name__)
    
    async def set_book(self, update: Update, context: ContextTypes.DEFAULT_TYPE):
        """Handle /setbook command."""
        # TODO: Implement set book command
        await update.message.reply_text("Set book command - Coming soon!")
    
    async def manage_league(self, update: Update, context: ContextTypes.DEFAULT_TYPE):
        """Handle /league command for admins."""
        # TODO: Implement league management
        await update.message.reply_text("League management - Coming soon!")
    
    async def view_members(self, update: Update, context: ContextTypes.DEFAULT_TYPE):
        """Handle /members command."""
        # TODO: Implement view members
        await update.message.reply_text("View members - Coming soon!")
    
    async def export_data(self, update: Update, context: ContextTypes.DEFAULT_TYPE):
        """Handle /export command."""
        # TODO: Implement data export
        await update.message.reply_text("Data export - Coming soon!")
    
    async def generate_report(self, update: Update, context: ContextTypes.DEFAULT_TYPE):
        """Handle /report command."""
        # TODO: Implement report generation
        await update.message.reply_text("Report generation - Coming soon!")
    
    async def view_users(self, update: Update, context: ContextTypes.DEFAULT_TYPE):
        """Handle /users command."""
        # TODO: Implement view users
        await update.message.reply_text("View users - Coming soon!")
    
    async def create_backup(self, update: Update, context: ContextTypes.DEFAULT_TYPE):
        """Handle /backup command."""
        # TODO: Implement backup creation
        await update.message.reply_text("Backup creation - Coming soon!")
    
    async def cleanup_data(self, update: Update, context: ContextTypes.DEFAULT_TYPE):
        """Handle /cleanup command."""
        # TODO: Implement data cleanup
        await update.message.reply_text("Data cleanup - Coming soon!")
