requirements.txt
# Core Bot Framework
python-telegram-bot==20.7

# Database and Data Processing
pandas>=2.2.0
numpy>=1.26.0
openpyxl>=3.1.2

# Google Services (for future export features)
google-auth>=2.23.4
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0

# Environment and Configuration
python-dotenv>=1.0.0
PyYAML>=6.0.1

# Date and Time Handling
pytz>=2023.3

# Image Processing (for achievements and certificates)
Pillow>=10.0.1

# HTTP Requests
requests>=2.31.0

# Async Support
aiofiles>=23.2.1

# Development and Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
flake8>=6.1.0
